+++ blue-ui-core/src/main/java/blue/soundObject/editor/PatternEditor.java
@@ -20,9 +20,14 @@
 
 package blue.soundObject.editor;
 
+import blue.BlueSystem;
+import blue.gui.ExceptionDialog;
+import blue.gui.InfoDialog;
 import blue.plugin.ScoreObjectEditorPlugin;
 import blue.score.ScoreObject;
+import blue.soundObject.NoteList;
 import blue.soundObject.PatternObject;
+import blue.soundObject.SoundObject;
 import blue.soundObject.editor.pattern.PatternCanvas;
 import blue.soundObject.editor.pattern.PatternLayerEditPanel;
 import blue.soundObject.editor.pattern.PatternObjectPropertiesPanel;
@@ -36,11 +41,13 @@ import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.MouseWheelEvent;
 import java.awt.event.MouseWheelListener;
+import javax.swing.JButton;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
 import javax.swing.JSplitPane;
 import javax.swing.JToggleButton;
 import javax.swing.ScrollPaneConstants;
+import javax.swing.SwingUtilities;
 import javax.swing.event.ListSelectionEvent;
 import javax.swing.event.ListSelectionListener;
 
@@ -81,8 +88,16 @@ public class PatternEditor extends ScoreObjectEditor {
         setTimeButton.setSelectedIcon(IconFactory.getRightArrowIcon());
         setTimeButton.setFocusable(false);
 
+        
+        JButton testButton = new JButton(BlueSystem.getString("common.test"));
+        testButton.addActionListener((ae) -> testSoundObject());
+        JPanel rightPanel = new JPanel(new BorderLayout());
+        rightPanel.add(testButton, BorderLayout.NORTH);
+        rightPanel.add(props, BorderLayout.CENTER);
+        
+        
         setTimeButton.addActionListener((ActionEvent e) -> {
-            props.setVisible(!props.isVisible());
+            rightPanel.setVisible(!rightPanel.isVisible());
         });
 
         final JScrollPane scroll = new JScrollPane(canvas);
@@ -102,14 +117,14 @@ public class PatternEditor extends ScoreObjectEditor {
 
         JPanel topPanel = new JPanel(new BorderLayout());
         topPanel.add(topSplitPane, BorderLayout.CENTER);
-        topPanel.add(props, BorderLayout.EAST);
+        topPanel.add(rightPanel, BorderLayout.EAST);
 
         JSplitPane mainSplitPane = new JSplitPane(JSplitPane.VERTICAL_SPLIT);
         mainSplitPane.add(topPanel);
         mainSplitPane.add(patternScore);
         mainSplitPane.setDividerLocation(0.8d);
 
-        props.setVisible(false);
+        rightPanel.setVisible(false);
 
         this.add(mainSplitPane, BorderLayout.CENTER);
 
@@ -141,6 +156,26 @@ public class PatternEditor extends ScoreObjectEditor {
         timeBar.setPatternObject(p);
         props.setPatternObject(p);
 
+    }
+    
+    public final void testSoundObject() {
+        if (this.patternObj == null) {
+            return;
+        }
+
+        NoteList notes = null;
+
+        try {
+            notes = patternObj.generateForCSD(null, 0.0f, -1.0f);
+        } catch (Exception e) {
+            ExceptionDialog.showExceptionDialog(SwingUtilities.getRoot(this), e);
+        }
+
+        if (notes != null) {
+            InfoDialog.showInformationDialog(SwingUtilities.getRoot(this),
+                    notes.toString(), BlueSystem
+                            .getString("soundObject.generatedScore"));
+        }
     }
 
     public static void main(String[] args) {